+++ CHANGELOG.md
@@ -6,6 +6,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.5.0] - 2026-06-24
+
 ### Added
 
 - **Explicit "Run in background" button + reliable Ubuntu tray (#9).** The